🚚 Rename: 컬럼타입 파일명 오타 수정
