+++ src/features/cards/components/card-list/cardList.types.ts
@@ -1,4 +1,4 @@
-import type { Column } from '@/features/columns/types/cloumn.types';
+import type { Column } from '@/features/columns/types/column.types';
 
 export type CardListProps = {
   column: Column;
