+++ src/features/columns/apis/columns.ts
@@ -4,7 +4,7 @@ import type {
   CreateColumnRequest,
   UpdateColumnRequest,
 } from '@/features/columns/apis/columns.types';
-import type { Column } from '@/features/columns/types/cloumn.types';
+import type { Column } from '@/features/columns/types/column.types';
 import { get, post, del, put } from '@/shared/apis/fetchInstance';
 
 /**
