+++ src/features/columns/contexts/columnContext.ts
@@ -1,4 +1,4 @@
-import type { Column } from '@/features/columns/types/cloumn.types';
+import type { Column } from '@/features/columns/types/column.types';
 import { createContext } from 'react';
 
 export const ColumnContext = createContext<Column | null>(null);
