+++ src/features/columns/contexts/columnListContext.ts
@@ -1,4 +1,4 @@
 import { createContext } from 'react';
-import type { Column } from '@/features/columns/types/cloumn.types';
+import type { Column } from '@/features/columns/types/column.types';
 
 export const ColumnListContext = createContext<Column[]>([]);
